src/python/pyproject.toml
[tool.poetry]
name = "time-server"
version = "0.1.0"
description = "Time Server Django REST API"
authors = ["tcronin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8.1"
django = "^4.2"
djangorestframework = "^3.14"
django-cors-headers = "^4.0"
renogy-modbus-lib-python = "*"
drf-spectacular = "^0.26.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
isort = "^5.12"
flake8 = "^6.0"
pytest = "^7.4"
pytest-django = "^4.5"
pytest-cov = "^4.1"
pytest-asyncio = "^0.21.0"
pytest-html = "^4.1.1"

[tool.poetry.scripts]
# Development tasks
dev-server = "scripts.dev:start_dev_server"
test = "scripts.dev:run_tests"
test-cov = "scripts.dev:run_tests_with_coverage"
lint = "scripts.dev:run_linting"
format = "scripts.dev:run_formatting"
clean = "scripts.dev:clean_project"
help = "scripts.dev:show_help"

# Docker tasks
docker-build = "scripts.docker:build_image"
docker-run = "scripts.docker:run_container"
docker-stop = "scripts.docker:stop_container"
docker-logs = "scripts.docker:show_logs"

# Database tasks
db-migrate = "scripts.db:migrate"
db-reset = "scripts.db:reset_database"
db-seed = "scripts.db:seed_database"

# Deployment tasks
deploy-staging = "scripts.deploy:deploy_staging"
deploy-prod = "scripts.deploy:deploy_production"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | tests
  | reports
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = "time_server"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = ["tests/*", "reports/*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "time_server.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = [
    "--import-mode=importlib",
    "--html=reports/test.html",
    "--junit-xml=reports/test.xml",
    "--cov=apps",
    "--cov=api.urls",
    "--cov-report=html:reports/htmlcov",
    "--cov-report=term-missing",
    "--cov-report=xml:reports/coverage.xml"
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
    "build",
    "dist",
    "tests",
    "reports",
]
per-file-ignores = [
    "time_server/settings.py:E501",
]
